implementando testes unitários no UsuarioService

+++ src/main/java/com/ecommerce/ecommerceGenerico/service/UsuarioService.java
@@ -4,12 +4,12 @@ import java.time.OffsetDateTime;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import com.ecommerce.ecommerceGenerico.entityes.CarrinhoEntity;
 import com.ecommerce.ecommerceGenerico.entityes.UsuarioEntity;
 import com.ecommerce.ecommerceGenerico.exception.excesoesPersonalizadas.UsuarioException;
 import com.ecommerce.ecommerceGenerico.repository.UserRepository;
+
 
 @Service
 public class UsuarioService {
@@ -22,7 +22,11 @@ public class UsuarioService {
 	}
 	
 	private boolean validaSenha(String senha) {
-		return false;
+		if(senha.length() < 8) {
+			return false;
+		}
+		
+		return true;
 	}
 	
 	public void adicionaUsuario(String nome, String email, String senha) {
@@ -65,6 +69,10 @@ public class UsuarioService {
 	}
 
 	public String removeUsuario(Long userId) {
+		
+		if(userId == null) {
+			throw new IllegalArgumentException("id não pode ser nulo");
+		}
 		
 		userRepository.deleteById(userId);
